move actions and reducers to the pods
