--- actions/index.js
@@ -1,3 +1,0 @@
-export {categoryClues, randomCategories, updateCategoryClue, updateCategoryClues } from './Categories';
-export {randomClue, updateClues} from './Clues';
-export {requestData, receiveData} from './Data';
