+++ app/actions/index.js
@@ -1,0 +1,3 @@
+export {categoryClues, randomCategories, updateCategoryClue, updateCategoryClues } from '../board/actions/Categories';
+export {randomClue, updateClues} from '../board/actions/Clues';
+export {requestData, receiveData} from './Data';
