+++ app/board/reducers/Categories.js
